requirements.txt
Django>=5.0,<6.1
psycopg2-binary>=2.9.9
python-decouple>=3.8
django-modeltranslation>=0.18.11
django-rosetta>=0.10.0
django-compressor>=4.4
whitenoise>=6.6.0
gunicorn>=21.2.0
Pillow>=10.1.0
redis>=5.0.1
django-redis>=5.4.0
django-extensions>=3.2.3
django-htmx>=1.17.2
django-browser-reload>=1.12.1
django-tailwind>=3.8.0
